+++ legacy_packages/chains/chains/185.ts
@@ -34,7 +34,7 @@ export default {
     "https://185.rpc.thirdweb.com/${THIRDWEB_API_KEY}",
     "https://rpc.mintchain.io",
     "https://global.rpc.mintchain.io",
-    "https://asia.rpc.minchain.io"
+    "https://asia.rpc.mintchain.io"
   ],
   "shortName": "mint",
   "slug": "mint",