+++ packages/thirdweb/src/exports/wallets/in-app.ts
@@ -9,3 +9,5 @@ export {
 } from "../../wallets/in-app/core/authentication/index.js";
 
 export { hasStoredPasskey } from "../../wallets/in-app/web/lib/auth/passkeys.js";
+
+export { socialIcons } from "../../react/web/wallets/in-app/socialIcons.js";